Change the password salt to drive serial number
After the issue with Samsung not letting me read the authority table I've given up on being able to store something on the drive because it would have to be accessable to an unauthenticated user and that doesn't seem very likely to be consistent across vendors.  The use of the drive serial number is better than a fixed salt in that it should prevent attacks with a prebuilt dictionary. I also can't find a disk UUID that is consistent across OS's