disable hash cmdline parameter
This patch adds a new optional -n parameter which disables
the password hashing function, causing passwords to be
transmitted to the drive in clear text.

This is useful for situations where hashed passwords are not
required, and it is desirable to save CPU time.

On a 2.5GHz CPU this saved over 100ms of execution time per command.