Clean up return values and exit statuses in sedutil.
Previously, the code for Enterprise encryption would often return
a non-specific failure (0xff) code which was then used as an exit
value.

In order to implement any sort of error recovery for specific
failure cases outside of sedutil, we need to have a clear
indication for the calling program what the failure is.

So, this replaces all of the instances where sedutil returns 0xff
in the case of failure for Enterprise drives with the actual
failure status.

This will allow the caller to, for instance, try resetting the LUN
if sedutil exits with status 0x3 (SP_BUSY).

sedutil/Common/DtaDevEnterprise.cpp:
	Follow the general pattern from DtaDevOpal.ccp, and save
	return codes in a lastRC variable so they can be passed
	back to the caller, rather than just returning 0xff on
	failure.

sedutil/Common/DtaDevEnterprise.h:
	Change the prototypes for getMaxRanges() and
	getMaxRangesOpal() to return status and take the max ranges
	variable on the command line.

sedutil/Common/DtaSession.cpp:
	In DtaSession::start(), allocate a local lastRC.  Print out
	the return code in the event that starting the session
	fails.

sedutil/Common/sedutil.cpp:
	If --printDefaultPassword fails, exit with the proper
	return code instead of exiting with 0.  This will tell
	the caller why it failed, and therefore whether there
	is something he can do to recover.