Change sedutil-cli to allow resetting passwords back to the default.
The --revertTPer and --revertNoErase options don't seem to work on
Enterprise drives, and so far the only way I've found to really
revert a drive back to factory defaults is using the PSID.  And the
PSID isn't available on the HGST drives I have at least.  (The
Seagate drives do have it.)

So, this allows setting the password back to the default.

Common/DtaDevEnterprise.cpp:
	In setPassword(), get the default password if either the
	current or new password is NULL/NUL.  Set the new password
	to the MSID without hashing if the user wants to go back to
	the default.