Don't fail initial setup if we can't set all BandMaster passwords.
The Seagate ST6000NM0114 drives seem to deny access to BandMaster2
and up.  It isn't clear why, but it might be that you have to
configure BandMaster1 first before they'll allow access to
subsequent bands.

So, in initLSPUsers(), only return an error if we can't set
BandMaster0.  The others are more optional.

Common/DtaDevEnterprise.cpp:
	In initLSPUsers(), don't return an error if we fail to set
	the password for BandMasterN where N > 0.