Add line break
In a  previous commit, this was edited to break up lines. A mistake was made, neglecting the line break characters. Line 35 is not a switch statement, but a define statement. As such, the line needs to be broken up as a string. I tested this, but let me know if I made a mistake.