LTS update for Buildroot and Linux Kernel
This pull request updates:
  Buildroot2 to the 2019.02.6 (latest LTS) target
  Linux Kernel to the 4.14.146 (latest version in 4.14.x LTS tree) version
  Other incidental updates along the way (binutils 2.30.x, GCC 8.3.x)

A longer explanation of why this update and how to do further updates is
in README_LTS.txt.

This update addresses several issues with USB input handling in linuxpba,
support for AMD Ryzen 2xxx and 3xxx chips, and other minor items.

This pull request is explicitly under the GPL 3.0 license, as is implicitly
required by Github's latest "inbound = outbound" license updates as discussed:
  https://ben.balter.com/2018/01/02/why-you-probably-shouldnt-add-a-cla-to-your-open-source-project/
  https://help.github.com/articles/github-terms-of-service/#6-contributions-under-repository-license
As always, differing opinions are encouraged; c.f.
  https://writing.kemitchell.com/2018/01/06/CLAs-Are-Not-a-Sham.html

This pull request should apply as-is for most current "unofficial" source
trees as it only affects certain files, which should not conflict with most
other known patch sets:

	new file:   README_LTS.txt
	modified:   images/buildpbaroot
	modified:   images/buildroot/32bit/.config
	modified:   images/buildroot/32bit/kernel.config
	new file:   images/buildroot/32bit/local.mk
	modified:   images/buildroot/64bit/.config
	modified:   images/buildroot/64bit/kernel.config
	new file:   images/buildroot/64bit/local.mk
	modified:   images/conf