ISO file builder for Rescue64 (UEFI) + GRUB
This is a standalone set of files in images/isoFiles/ that adds a capability to take any
Rescue64 (nominally UEFI) .img.gz file created by the normal "./buildrescue Rescue64" and
turn it into a UEFI-bootable ISO image - no CSM required.  The "xorriso" utility is used to
remaster the Rescue64 .img file into an ISO.

Primary reason for this was to allow Rescue images on CD for environments that avoid USB
and/or like to have dedicated media that can't get overwritten.

Secondary purpose was to test the ability of GRUB v2.04 to support Trusted Boot with built-in
"shim_lock" modules etc. and without requiring the user/sysadmin to perform additional
cryptographic/cert signing.  That part is untested.

This set of scripts is tested against DTA SEDutil 1.15.1 and makes no changes to the normal
build process, but simply adds a "cd" then two additional steps after the Rescue64 image is built:
    cd isoFiles
    ./buildgrub
    ./buildiso