linux/DtaDevOS.cpp: improve robustness of drive type auto-detection
DtaDevOS::init() previously detected the drive type by checking whether
the user-specified device node path starts with "/dev/nvme" or "/dev/s".
This is fragile and breaks mysteriously in cases such as when the user
specifies a relative path to a device node or when the user specifies a
device through a udev-managed symbolic link in /dev/disk/by-id/ et al.

This commit reimplements drive type auto-detection using a more robust
algorithm that stats the specified device node and treats it as a SATA
drive if the device node's major number is one of the SCSI_DISK*_MAJOR
numbers or as an NVMe drive if the device is bound to the "nvme" kernel
driver. (The latter check requires sysfs to be mounted at /sys.)