Shuffle GetPassPhrase{.h, .cpp} in makefile.am
Moving GetPassPhrase.o down into SEDUTIL_LINUX_CODE, since we need it
for the securemode/password-setting work.

This is fine since AFAICT linuxpba_SOURCES eventually gets consumed in
conjunction with SEDUTIL_LINUX_CODE, so that binary will get access to
the TU.

Keeping this as seperate commit since it might be a little bit subtle /
weird, depending on what Make is doing behind the scenes.

Signed-off-by: Howard McLauchlan <hmclauchlan@fb.com>