Generate Random bytes feature
TCG Storage Architecture Core Specification allows random byte generation
as one of the cryptographic operation. This check-in implements printRandomBytes
function for Opal drives. It should work for Enterprise drive as well. However,
at the time of this writing, we could not procure Enterprise version to validate.

It is an error to have range outside of 1 to 32.